+++ src/Contracts/KnownProviders.cs
@@ -1,0 +1,13 @@
+﻿namespace DialToolsForVS
+{
+    public enum KnownProviders
+    {
+        Bookmarks,
+        Debug,
+        Editor,
+        Errors,
+        Navigation,
+        Scroll,
+        Zoom
+    }
+}