+++ src/Providers/Errors/ErrorsControllerProvider.cs
@@ -3,7 +3,7 @@
     [DialControllerProvider(order: 6)]
     internal class ErrorsControllerProvider : IDialControllerProvider
     {
-        public const string Moniker = "Errors";
+        public static string Moniker = KnownProviders.Errors.ToString();
 
         public IDialController TryCreateController(IDialControllerHost host)
         {