+++ src/Providers/Scroll/ScrollControllerProvider.cs
@@ -1,15 +1,22 @@
-﻿namespace DialToolsForVS
+﻿using System.ComponentModel.Composition;
+using Microsoft.VisualStudio.Language.Intellisense;
+
+namespace DialToolsForVS
 {
     [DialControllerProvider(order: 1)]
     internal class ScrollControllerProvider : IDialControllerProvider
     {
-        public const string Moniker = "Scroll";
+        public static string Moniker = KnownProviders.Scroll.ToString();
+
+        [Import]
+        private ICompletionBroker CompletionBroker { get; set; }
+
         public IDialController TryCreateController(IDialControllerHost host)
         {
             string iconFilePath = VsHelpers.GetFileInVsix("Providers\\Scroll\\icon.png");
             host.AddMenuItem(Moniker, iconFilePath);
 
-            return new ScrollController();
+            return new ScrollController(CompletionBroker);
         }
     }
 }